Move custom error to separate module.
New module is in shared/src/errors.rs, this provides better code readability and less complexity per single file.
