+++ shared/src/account_balance.rs
@@ -1,3 +1,4 @@
+use crate::errors::RequestOrParseError;
 use crate::v0_0_5::account_balance::AccountBalanceResponseV0_0_5;
 use crate::v0_0_6::account_balance::AccountBalanceResponseV0_0_6;
 use clap::Parser;
@@ -6,37 +7,6 @@ use reqwest::Client;
 use tracing::info;
 use url::Url;
 
-use std::error::Error as StdError;
-use std::fmt;
-
-#[derive(Debug)]
-pub enum RequestOrParseError {
-    Reqwest(reqwest::Error),
-    Url(url::ParseError),
-}
-
-impl fmt::Display for RequestOrParseError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            RequestOrParseError::Reqwest(e) => write!(f, "{}", e),
-            RequestOrParseError::Url(e) => write!(f, "{}", e),
-        }
-    }
-}
-
-impl StdError for RequestOrParseError {}
-
-impl From<reqwest::Error> for RequestOrParseError {
-    fn from(err: reqwest::Error) -> RequestOrParseError {
-        RequestOrParseError::Reqwest(err)
-    }
-}
-
-impl From<url::ParseError> for RequestOrParseError {
-    fn from(err: url::ParseError) -> RequestOrParseError {
-        RequestOrParseError::Url(err)
-    }
-}
 #[derive(Parser, Debug, Clone)]
 pub enum Version {
     V0_0_6,
