+++ shared/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod account_balance;
+pub mod errors;
 pub use v0_0_5;
 pub use v0_0_6;
